Rename ff_satellite
