+++ s2e-ff/src/simulation/case/ff_case.hpp
@@ -3,7 +3,7 @@
 #include <simulation/case/simulation_case.hpp>
 #include <simulation/multiple_spacecraft/relative_information.hpp>
 
-#include "../spacecraft/FfSat.hpp"
+#include "../spacecraft/ff_satellite.hpp"
 
 class FfCase : public SimulationCase {
  public:
