+++ s2e-ff/src/simulation/spacecraft/ff_satellite.hpp
